+++ tpespecial/usuario-service/src/main/java/com/example/usuario_service/service/impl/CuentaServiceImpl.java
@@ -1,6 +1,8 @@
 package com.example.usuario_service.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import com.example.usuario_service.model.dto.CuentaRequest;
 import com.example.usuario_service.model.dto.CuentaResponse;
@@ -33,39 +35,47 @@ public class CuentaServiceImpl implements CuentaService {
 
     //Agregar una cuenta
     @Override
-    public void addCuenta(CuentaRequest cuenta) {
+    public ResponseEntity<String> addCuenta(CuentaRequest cuenta) {
        Cuenta c = new Cuenta(cuenta.getMercadoPagoId());
        cuentaRepository.save(c);
+       return ResponseEntity.status(HttpStatus.CREATED).body("Cuenta creada exitosamente.");
     }
 
     //Actualizar una cuenta
     @Override
-    public void putCuenta(Long id, CuentaRequest cuenta) {
-        Cuenta c = cuentaRepository.findById(id)
-            .orElseThrow(() -> new NoSuchElementException("Cuenta con id " + id + " no encontrado"));
-        c.setMercadoPagoId(cuenta.getMercadoPagoId());
-        c.setCreditos(cuenta.getCreditos());
-        c.setFechaDeCreacion(cuenta.getFechaDeCreacion());
-        c.setHabilitada(cuenta.isHabilitada());
-        cuentaRepository.save(c);
+    public ResponseEntity<String> putCuenta(Long id, CuentaRequest cuenta) {
+        Cuenta c = cuentaRepository.findById(id).orElse(null);
+        if(c != null){
+            c.setMercadoPagoId(cuenta.getMercadoPagoId());
+            c.setCreditos(cuenta.getCreditos());
+            c.setFechaDeCreacion(cuenta.getFechaDeCreacion());
+            c.setHabilitada(cuenta.isHabilitada());
+            cuentaRepository.save(c);
+            return ResponseEntity.status(HttpStatus.OK).body("Cuenta actualizada exitosamente.");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("La cuenta no existe.");
     }
 
 
     //Eliminar una cuenta
     @Override
-    public void deleteCuenta(Long id) {
-        Cuenta c = cuentaRepository.findById(id)
-            .orElseThrow(() -> new NoSuchElementException("Cuenta con id " + id + " no encontrado"));
-        
-            cuentaRepository.delete(c);
+    public ResponseEntity<String> deleteCuenta(Long id) {
+        Cuenta c = cuentaRepository.findById(id).orElse(null);
+            if(c != null){
+                cuentaRepository.delete(c);
+                return ResponseEntity.status(HttpStatus.OK).body("Cuenta eliminada exitosamente.");
+            }
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("La cuenta no existe.");
     }
 
     @Override
-    public void cambiarEstadoCuenta(Long id, boolean habilitada) {
+    public ResponseEntity<String> cambiarEstadoCuenta(Long id, boolean habilitada) {
         Cuenta cuenta = cuentaRepository.findById(id).orElse(null);
         if(cuenta != null){
             cuenta.setHabilitada(habilitada);
+            return ResponseEntity.status(HttpStatus.OK).body("Estado de cuenta modificado exitosamente.");
         }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("La cuenta no existe.");
     }
  
 }
